runner/requirements.txt
mysql-connector-python==8.0.27
requests==2.26.0